+++ CHANGELOG.md
@@ -43,3 +43,4 @@ All notable changes to the "ftlXml" extension will be documented in this file.
 - remove duplicate elements in autocomplete
 - allow a blueprint list to contain anything and don't warn about type
 - fix an issue where blueprint loops would break the parser
+- when a workspace folder is added or removed update the data